p_menu/requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==4.0.5
django-allauth==0.48.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-extra-fields==3.4.0
drf-spectacular==0.21.2
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.16.0
Markdown==3.3.7
oauthlib==3.2.1
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
pycparser==2.21
PyJWT==2.5.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
types-cryptography==3.3.23
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.0
